Fix problem where server would not send connection ID to client due to changes from int to short

+++ UDP Server/src/model/Player.java
@@ -19,7 +19,7 @@ public class Player extends GameObject{
 		playerID = id;
 	}
 	
-	public int getID() {
+	public short getID() {
 		return playerID;
 	}
